update elements to Flow 4.0
