+++ Classes/Generator/PdfGeneratorInterface.php
@@ -1,8 +1,8 @@
 <?php
-namespace Famelo\PDF\View;
+namespace Famelo\PDF\Generator;
 
 /*                                                                        *
- * This script belongs to the FLOW3 package "Famelo.Messaging".           *
+ * This script belongs to the FLOW3 package "Famelo.PDF".                 *
  *                                                                        *
  * It is free software; you can redistribute it and/or modify it under    *
  * the terms of the GNU Lesser General Public License, either version 3   *
@@ -11,23 +11,23 @@ namespace Famelo\PDF\View;
  * The TYPO3 project - inspiring people to share!                         *
  *                                                                        */
 
-use TYPO3\Flow\Annotations as Flow;
+use Neos\Flow\Annotations as Flow;
 
-/**
- * A standalone template view.
- * Helpful if you want to use Fluid separately from MVC
- * E.g. to generate template based emails.
- *
- * @api
- */
-class StandaloneView extends \TYPO3\Fluid\View\StandaloneView {
-	public function initializeObject() {
-		parent::initializeObject();
+interface PdfGeneratorInterface {
 
-		$this->request->setFormat('html');
-	}
+	public function setFormat($format);
 
-	public function getViewHelperVariableContainer() {
-		return $this->baseRenderingContext->getViewHelperVariableContainer();
-	}
+	public function setHeader($content);
+
+	public function setFooter($content);
+
+	public function setOption($name, $value);
+
+	public function sendPdf($content, $filename = NULL);
+
+	public function downloadPdf($content, $filename = NULL);
+
+	public function savePdf($content, $filename);
+
 }
+?>
