+++ Classes/Package.php
@@ -11,8 +11,8 @@ namespace Famelo\PDF;
  * The TYPO3 project - inspiring people to share!                         *
  *                                                                        */
 
-use TYPO3\Flow\Core\Bootstrap;
-use TYPO3\Flow\Package\Package as BasePackage;
+use Neos\Flow\Core\Bootstrap;
+use Neos\Flow\Package\Package as BasePackage;
 
 /**
  * The TYPO3 Flow Package
@@ -22,7 +22,7 @@ class Package extends BasePackage {
 	/**
 	 * Invokes custom PHP code directly after the package manager has been initialized.
 	 *
-	 * @param \TYPO3\Flow\Core\Bootstrap $bootstrap The current bootstrap
+	 * @param \Neos\Flow\Core\Bootstrap $bootstrap The current bootstrap
 	 * @return void
 	 */
 	public function boot(Bootstrap $bootstrap) {
