+++ Classes/View/StandaloneView.php
@@ -1,8 +1,8 @@
 <?php
-namespace Famelo\PDF\Error;
+namespace Famelo\PDF\View;
 
 /*                                                                        *
- * This script belongs to the FLOW3 package "Famelo.PDF".                 *
+ * This script belongs to the FLOW3 package "Famelo.Messaging".           *
  *                                                                        *
  * It is free software; you can redistribute it and/or modify it under    *
  * the terms of the GNU Lesser General Public License, either version 3   *
@@ -11,11 +11,23 @@ namespace Famelo\PDF\Error;
  * The TYPO3 project - inspiring people to share!                         *
  *                                                                        */
 
+use Neos\Flow\Annotations as Flow;
+
 /**
- * "Unknown Class" Exception
+ * A standalone template view.
+ * Helpful if you want to use Fluid separately from MVC
+ * E.g. to generate template based emails.
  *
  * @api
  */
-class UnknownGeneratorOptionException extends \TYPO3\Flow\Error\Exception {
+class StandaloneView extends \Neos\FluidAdaptor\View\StandaloneView {
+	public function initializeObject() {
+		parent::initializeObject();
 
+		$this->request->setFormat('html');
+	}
+
+	public function getViewHelperVariableContainer() {
+		return $this->baseRenderingContext->getViewHelperVariableContainer();
+	}
 }
