+++ Classes/ViewHelpers/FooterViewHelper.php
@@ -1,5 +1,5 @@
 <?php
-namespace Famelo\PDF\ViewHelpers\Fpdi;
+namespace Famelo\PDF\ViewHelpers;
 
 /*                                                                        *
  * This script belongs to the TYPO3 Flow package "Famelo.Pdf".            *
@@ -11,13 +11,12 @@ namespace Famelo\PDF\ViewHelpers\Fpdi;
  * The TYPO3 project - inspiring people to share!                         *
  *                                                                        */
 
-use TYPO3\Fluid\Core\ViewHelper\AbstractViewHelper;
+use Neos\FluidAdaptor\Core\ViewHelper\AbstractViewHelper;
 
 /**
  * @api
  */
-class DefaultsViewHelper extends AbstractViewHelper {
-
+class FooterViewHelper extends AbstractViewHelper {
 	
 	/**
 	 * NOTE: This property has been introduced via code migration to ensure backwards-compatibility.
@@ -25,19 +24,6 @@ class DefaultsViewHelper extends AbstractViewHelper {
 	 * @var boolean
 	 */
 	protected $escapeOutput = FALSE;
-
-	/**
-	 * Constructor
-	 *
-	 * @api
-	 */
-	public function __construct() {
-		$this->registerArgument('font', 'string', 'font name', FALSE);
-		$this->registerArgument('font-size', 'integer', 'font size', FALSE);
-		$this->registerArgument('font-weight', 'string', 'font weight', FALSE);
-		$this->registerArgument('color', 'string', 'font color', FALSE);
-	}
-
 	/**
 	 * This tag will not be rendered at all.
 	 *
@@ -45,6 +31,7 @@ class DefaultsViewHelper extends AbstractViewHelper {
 	 * @api
 	 */
 	public function render() {
-		$this->viewHelperVariableContainer->add('Famelo\Pdf\ViewHelpers\Fpdi\DefaultsViewHelper', 'defaults', $this->arguments);
+		$content = $this->renderChildren();
+		$this->viewHelperVariableContainer->add('Famelo\Pdf\ViewHelpers\FooterViewHelper', 'footer', $content);
 	}
 }
