+++ Classes/ViewHelpers/Fpdi/TemplateViewHelper.php
@@ -1,5 +1,5 @@
 <?php
-namespace Famelo\PDF\ViewHelpers;
+namespace Famelo\PDF\ViewHelpers\Fpdi;
 
 /*                                                                        *
  * This script belongs to the TYPO3 Flow package "Famelo.Pdf".            *
@@ -11,12 +11,13 @@ namespace Famelo\PDF\ViewHelpers;
  * The TYPO3 project - inspiring people to share!                         *
  *                                                                        */
 
-use TYPO3\Fluid\Core\ViewHelper\AbstractViewHelper;
+use Neos\FluidAdaptor\Core\ViewHelper\AbstractViewHelper;
 
 /**
  * @api
  */
-class HeaderViewHelper extends AbstractViewHelper {
+class TemplateViewHelper extends AbstractViewHelper {
+
 	
 	/**
 	 * NOTE: This property has been introduced via code migration to ensure backwards-compatibility.
@@ -24,6 +25,16 @@ class HeaderViewHelper extends AbstractViewHelper {
 	 * @var boolean
 	 */
 	protected $escapeOutput = FALSE;
+
+	/**
+	 * Constructor
+	 *
+	 * @api
+	 */
+	public function __construct() {
+		$this->registerArgument('path', 'string', 'template path', TRUE);
+	}
+
 	/**
 	 * This tag will not be rendered at all.
 	 *
@@ -31,7 +42,6 @@ class HeaderViewHelper extends AbstractViewHelper {
 	 * @api
 	 */
 	public function render() {
-		$header = $this->renderChildren();
-		$this->viewHelperVariableContainer->add('Famelo\Pdf\ViewHelpers\HeaderViewHelper', 'header', $header);
+		$this->viewHelperVariableContainer->add('Famelo\Pdf\ViewHelpers\Fpdi\TemplateViewHelper', 'template', $this->arguments['path']);
 	}
 }
