+++ Classes/ViewHelpers/Fpdi/TextViewHelper.php
@@ -11,7 +11,7 @@ namespace Famelo\PDF\ViewHelpers\Fpdi;
  * The TYPO3 project - inspiring people to share!                         *
  *                                                                        */
 
-use TYPO3\Fluid\Core\ViewHelper\AbstractViewHelper;
+use Neos\FluidAdaptor\Core\ViewHelper\AbstractViewHelper;
 
 /**
  * @api
