+++ Configuration/Objects.yaml
@@ -1,4 +1,4 @@
 Famelo\PDF\View\StandaloneView:
   properties:
     renderingContext:
-      object: TYPO3\Fluid\Core\Rendering\RenderingContext
+      object: Neos\Fluid\Core\Rendering\RenderingContext
