+++ composer.json
@@ -1,20 +1,53 @@
 {
-  "name":"famelo\/pdf",
-  "type":"typo3-flow-package",
-  "version": "2.0.0",
-  "description":"Provides a quick and simple way to generate a PDF from a Fluid Template through the MPDF library",
-  "require": {
-    "typo3\/flow": "*",
-    "mpdf\/mpdf": "*"
-  },
-  "suggest": {
-    "knplabs/knp-snappy":  "wkhtmltopdf wrapper to use the WebkitGenerator",
-    "itbz/fpdi": "fpdi library to use the FpdiGenerator"
-  },
-  "minimum-stability": "dev",
-  "autoload": {
-    "psr-0": {
-      "Famelo\\PDF":"Classes"
+    "name": "famelo/pdf",
+    "type": "typo3-flow-package",
+    "version": "2.0.0",
+    "description": "Provides a quick and simple way to generate a PDF from a Fluid Template through the MPDF library",
+    "require": {
+        "neos/flow": "~4.0.0",
+        "mpdf/mpdf": "*"
+    },
+    "suggest": {
+        "knplabs/knp-snappy": "wkhtmltopdf wrapper to use the WebkitGenerator",
+        "itbz/fpdi": "fpdi library to use the FpdiGenerator"
+    },
+    "minimum-stability": "dev",
+    "autoload": {
+        "psr-4": {
+            "Famelo\\PDF\\": "Classes"
+        }
+    },
+    "extra": {
+        "applied-flow-migrations": [
+            "TYPO3.Fluid-20150214130800",
+            "TYPO3.Fluid-20141121091700",
+            "TYPO3.Flow-20141113121400",
+            "TYPO3.Fluid-20141113120800",
+            "TYPO3.Flow-201405111147",
+            "TYPO3.Flow-201310031523",
+            "TYPO3.Flow-201212051340",
+            "TYPO3.Flow-201211151101",
+            "TYPO3.Flow-201209251426",
+            "TYPO3.FLOW3-201209201112",
+            "TYPO3.FLOW3-201206271128",
+            "TYPO3.FLOW3-201205292145",
+            "TYPO3.Fluid-201205031303",
+            "TYPO3.FLOW3-201201261636",
+            "TYPO3.Flow-20151113161300",
+            "TYPO3.Flow-20161115140400",
+            "TYPO3.Flow-20161115140430",
+            "Neos.Flow-20161124204700",
+            "Neos.Flow-20161124204701",
+            "Neos.Flow-20161124224015",
+            "Neos.Party-20161124225257",
+            "Neos.Eel-20161124230101",
+            "Neos.Imagine-20161124231742",
+            "Neos.Media-20161124233100",
+            "Neos.Flow-20161125124112",
+            "TYPO3.FluidAdaptor-20161130112935",
+            "Neos.Media-20161219094126",
+            "Neos.Flow-20170125103800",
+            "Neos.Flow-20170127183102"
+        ]
     }
-  }
 }
